+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 # v0.1.0
 
 * Adding a date selector.
+* Adding a preset selector.
+* Changed the `~` to `circa`.